Closes Issue 108:	Sickbeard can't create /var/lib/sickbeard on startup
The sickbeard install script uses 'adduser' to create the sickbeard user
and its home directory, /var/lib/sickbeard; as the sickbeard user
already exists, the 'adduser' command failed, and the /var... directory
was not created.

All install scripts suffer from this issue, and they should be fixed.

Before creating an user, delete it if it already exist, so adduser will
not fail creating its home directory



git-svn-id: https://svn.code.sf.net/p/alt-f/code/trunk@2134 cd80dc97-fa67-4d1f-9ead-3176a861659d