samba-small: new package
In this package (samba-small) libsmbcommon.so is created and used by all
samba binaries (smbd nmbd smbtree smbstatus swat smbpasswd), significaly
reducing its size. This is intended to be incorporated in the firmware,
given the low amount of available flash space. But this removes the
hability to load modules.

The samba package, by contrary, is intended to be a disk-instalable package, doesn't
use libsmbcommon, and can load external modules. It's binaries are huge.



git-svn-id: https://svn.code.sf.net/p/alt-f/code/trunk@2410 cd80dc97-fa67-4d1f-9ead-3176a861659d