Several years of changes since last commit, so don't expect a detailed commit log.
As usual, many bug have been fixed and small features have been added.

Noteworthy changes:

A new cross compiler toolchain and uclibc version (uclibc-ng) is used,
so some existing binaries might not run; all packages should to be rebuilt
and a new ipkg feed setup.

The build recipe has a new step.
   . exports board
   ./mktc.sh # build the cross toolchain -- only needs to be done once
   make

Most packages have been updated to their last versions (samba4 is standard),
and their download URLs updated (at the time of writing).

As the DNS-323/321 has not enought flash memory space to accomodate the updated
packages, a "dns-323-321-compat" pkg contains the missing pkgs and has to be
installed on disk to restore functionalities that users expect.

UBIFS is now used as the flash filesystem, meaning that packages can be
installed on flash memory instead of disk; you can remove packages that are
distributed in the firmware and you don't use or need, and install the ones
that you need.




git-svn-id: https://svn.code.sf.net/p/alt-f/code/trunk@3932 cd80dc97-fa67-4d1f-9ead-3176a861659d